Conversational-Chatbot/requirements.txt
langchain
python-dotenv
langchain_community
pypdf
langchain-text-splitters
chromadb
langchain_chroma
streamlit
langchain_groq
langchain_huggingface